fix(rt): use mills instead of secs as `build_time` when build standalone file.

+++ cli/lib/standalone/binary.rs
@@ -95,7 +95,7 @@ pub struct Metadata {
   pub unstable_config: UnstableConfig,
   pub otel_config: OtelConfig,
   pub vfs_case_sensitivity: FileSystemCaseSensitivity,
-  pub build_time: Option<u64>,
+  pub build_time: Option<u128>, // unit: mill
 }
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, Hash)]
